Move SubgraphTracker to a separate file. Not use unordered_ containers and delete hash function for Output<Node>. Cleanup.
