+++ src/common/transformations/src/transformations/mlir/convert.cpp
@@ -65,27 +65,7 @@
 
 #include "mlir_op.hpp"
 #include "convert_common.hpp"
-
-
-namespace std {
-
-template <typename T>
-size_t getHash(const T& val) {
-    return std::hash<T>()(val);
-}
-
-template <typename T, typename... Args>
-size_t getHash(const T& val, Args&&... args) {
-    return llvm::hash_combine(getHash(val), getHash(std::forward<Args>(args)...));
-}
-
-template <>
-struct hash<ov::Output<ov::Node>> final {
-    size_t operator()(const ov::Output<ov::Node>& out) const {
-        return getHash(out.get_node(), out.get_index());
-    }
-};
-}  // namespace std
+#include "subgraph_tracker.hpp"
 
 
 namespace {
@@ -122,7 +102,7 @@ SmallVector<mlir::Type> get_types_for_values(mlir::MLIRContext* context, const o
 class ConversionContext {
 public:
     using Convertor = std::function<void(ConversionContext&, NodePtr)>;
-    using NodeOutputMap = std::unordered_map<ov::Output<ov::Node>, mlir::Value>;
+    using NodeOutputMap = std::map<ov::Output<ov::Node>, mlir::Value>;
 
     static const std::map<ov::DiscreteTypeInfo, Convertor> convertors;
     mlir::MLIRContext* context;
@@ -255,198 +235,6 @@ mlir::OwningOpRef<mlir::ModuleOp> ngraph_to_mlir(MLIRContext* context,
     return module;
 }
 
-
-using InputVector = std::vector<ov::Input<ov::Node>>;
-
-
-struct Subgraph {
-    ov::NodeVector nodes;
-    ov::OutputVector inputs;
-    ov::OutputVector outputs;
-    std::vector<InputVector> output_consumers;
-
-    // Consumes other subgraph
-    void merge (Subgraph& other) {
-        nodes.insert(nodes.end(), other.nodes.begin(), other.nodes.end());
-    }
-};
-
-
-using SubgraphPtr = std::shared_ptr<Subgraph>;
-using SubgraphID = SymbolPtr;
-
-
-class SubgraphTracker {
-public:
-
-    // callback to finalize the subgraph when it is terminated
-    using Finalizer = std::function<void(SubgraphPtr)>;
-
-    SubgraphTracker(Finalizer finalizer): m_finalizer(finalizer) {}
-
-    void add_node (NodePtr node, bool belongs) {
-        // collect all subgraph ids that input nodes belong to and all dependencies
-        Dependencies input_subgraphs;
-        Dependencies input_dependencies;
-        for(auto input_value: node->input_values()) {
-            auto node = input_value.get_node_shared_ptr();
-            if(auto id = get_subgraph_id(node)) {
-                input_subgraphs.insert(ov::symbol::ancestor_of(id));
-            }
-            const auto& deps = get_dependencies(node);
-            for(auto dep: deps) {
-                input_dependencies.insert(ov::symbol::ancestor_of(dep));
-            }
-        }
-
-        if(belongs) {
-            // Below we refuse to merge subgraphs if all of them cannot merge to a single subgraph, this is rough because
-            // there are cases when part of the input subgraphs can consume the node and others will come as inputs -- TODO.
-            // TODO: leave only those input subgraphs that are not conflicting with other subgraphs nor with any dependencies
-            if(input_subgraphs.empty() || intersected(input_subgraphs, input_dependencies)) {   // no input subgraphs || cannot merge all due to cycles
-                try_terminate_subgraphs(input_subgraphs, node);
-
-                // start a new subgraph
-                auto subgraph_id = new_subgraph();
-                add_node_to_subgraph(node, subgraph_id);
-                set_subgraph_id(node, subgraph_id);
-                input_dependencies.insert(input_subgraphs.begin(), input_subgraphs.end());
-            } else {
-                auto merged_subgraph_id = std::accumulate(
-                    input_subgraphs.begin(),
-                    input_subgraphs.end(),
-                    *input_subgraphs.begin(),
-                    [this](SubgraphID a, SubgraphID b) {
-                        merge_subgraphs(a, b);
-                        return a;
-                    }
-                );
-                set_subgraph_id(node, merged_subgraph_id);
-                add_node_to_subgraph(node, merged_subgraph_id);
-            }
-
-        } else {
-            try_terminate_subgraphs(input_subgraphs, node);
-            set_subgraph_id(node, nullptr);
-            input_dependencies.insert(input_subgraphs.begin(), input_subgraphs.end());
-        }
-        set_dependencies(node, input_dependencies);
-    }
-
-    void finalize() {
-        for(auto subgraph_record: m_subgraphs) {
-            terminate_subgraph(subgraph_record.first);
-        }
-    }
-
-private:
-
-    std::unordered_map<SubgraphID, SubgraphPtr> m_subgraphs;
-    using Dependencies = std::unordered_set<SubgraphID>;
-    Finalizer m_finalizer;
-
-    // // Detects if `node` depends on a node from `subgraph` but goes via node that doesn't belongs to `subgraph`
-    // bool depends_via_break (NodePtr node, const Subgraph& subgraph);
-
-    SubgraphID new_subgraph() {
-        SubgraphID id = std::make_shared<ov::Symbol>();
-        m_subgraphs[id] = std::make_shared<Subgraph>();
-        return id;
-    }
-
-    void add_node_to_subgraph(NodePtr node, SubgraphID id) {
-        get_subgraph(id)->nodes.push_back(node);
-    }
-
-    void merge_subgraphs(SubgraphID id1, SubgraphID id2) {
-        id1 = ov::symbol::ancestor_of(id1);
-        id2 = ov::symbol::ancestor_of(id2);
-        if (id1 == id2) return;
-
-        auto subgraph1 = get_subgraph(id1);
-        auto subgraph2 = get_subgraph(id2);
-        subgraph1->merge(*subgraph2);
-        m_subgraphs.erase(id1);
-        m_subgraphs.erase(id2);
-        ov::symbol::set_equal(id1, id2);
-        id1 = ov::symbol::ancestor_of(id1);
-        m_subgraphs[id1] = subgraph1;
-    }
-
-    SubgraphPtr get_subgraph(SubgraphID id) {
-        return m_subgraphs.at(ov::symbol::ancestor_of(id));
-    }
-
-    // set/get all subgraph ids that contribute to a given node
-
-    const Dependencies& get_dependencies(NodePtr node) {
-        return node->get_rt_info().at("__subgraph_dependencies").as<Dependencies>();
-    }
-    void set_dependencies(NodePtr node, const Dependencies& dependencies) {
-        node->get_rt_info()["__subgraph_dependencies"] = dependencies;
-    }
-
-    // set/get subgraph id that a give node belongs to
-
-    SubgraphID get_subgraph_id(NodePtr node) {
-        auto id = node->get_rt_info().at("__subgraph_id").as<SubgraphID>();
-        if(id) {
-            id = ov::symbol::ancestor_of(id);
-        }
-        return id;
-    }
-
-    void set_subgraph_id(NodePtr node, SubgraphID id) {
-        node->get_rt_info()["__subgraph_id"] = id;
-    }
-
-    bool intersected(const Dependencies& a, const Dependencies& b) {
-        for(const auto& x: a) {
-            if(b.count(x))
-                return true;
-        }
-        return false;
-    }
-
-    void terminate_subgraph(SubgraphID id) {
-        id = ov::symbol::ancestor_of(id);
-        auto subgraph = get_subgraph(id);
-        // Build subgraph inputs and outputs
-        std::unordered_set<ov::Output<ov::Node>> inputs;
-        auto& outputs = subgraph->outputs;
-        auto& output_consumers = subgraph->output_consumers;
-        for(auto node: subgraph->nodes) {
-            for(auto input: node->input_values()) {
-                auto input_id = get_subgraph_id(input.get_node_shared_ptr());
-                if(!ov::symbol::are_equal(id, input_id)) {
-                    inputs.insert(input);
-                }
-            }
-            for(auto output: node->outputs()) {
-                const auto& consumers = output.get_target_inputs();
-                InputVector external_consumers;
-                for(auto consumer: consumers) {
-                    auto consumer_id = get_subgraph_id(consumer.get_node()->shared_from_this());
-                    if(!ov::symbol::are_equal(id, consumer_id)) {
-                        external_consumers.push_back(consumer);
-                    }
-                }
-                bool used_outside = !external_consumers.empty();
-                if(used_outside) {
-                    outputs.push_back(output);
-                    output_consumers.push_back(external_consumers);
-                }
-            }
-        }
-        subgraph->inputs.assign(inputs.begin(), inputs.end());
-        m_finalizer(subgraph);
-    }
-
-    void try_terminate_subgraphs(const Dependencies& subgraphs, NodePtr terminator) {
-        // TODO: Terminate subgraphs earlier when all terminating nodes are known
-        // TODO: try to merge subgraphs if they are being terminated simultaniously
-    }
-};
 
 // This pass find marked with a special flag group of nodes and collapse each group to a single MLIR function
 NodePtr ngraph_to_mlir_op(MLIRContext* context, SubgraphPtr subgraph) {
